Continue with saving
The thing here is getting the raw bytes into the db too.
Is passing the byte vec arround the right approach?
I want all the raw bytes because then i can keep adding fields,
and get them populated from the bytes,
in the future.
Let me try it and see.